Refactors Featured image and component files to featured/ folder, updates paths accordingly
