+++ src/components/portfolio/featured/Featured.jsx
@@ -1,0 +1,24 @@
+import React from 'react'
+import { FeaturedContentCard } from './FeaturedContentCard'
+import featuredContentData from '../../../data/featured-content'
+
+export const Featured = () => {
+  const featuredContentItems = featuredContentData.map(featuredContentitem => {
+    return (
+      <div key={featuredContentitem.id}>
+        <FeaturedContentCard featuredContentItem={featuredContentitem} />
+      </div>
+    )
+  })
+
+  return (
+    <>
+      <div
+        className="project-categories-wrapper 
+    project-categories-wrapper-featured"
+      >
+        {featuredContentItems}
+      </div>
+    </>
+  )
+}
